+++ test/cases/js-import-css-nested-deep/expected/js/pageA.2fcc41eb.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";console.log(">> Page A: ",{Foo:"Component Bar",Bar:"Component Foo"})})();