+++ test/cases/js-import-css-nested-deep/expected/js/pageB.af02ad73.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";console.log(">> Page B: ",{Foo:"Component Bar",Bar:"Component Foo"})})();