+++ test/cases/js-import-css-nested-deep/src/components/Bar.js
@@ -1,0 +1,5 @@
+import './Bar.css';
+
+const name = 'Component Bar';
+
+export default name;