+++ test/cases/js-import-css-nested-deep/src/components/Foo.js
@@ -1,0 +1,5 @@
+import './Foo.css';
+
+const name = 'Component Foo';
+
+export default name;