+++ test/cases/js-import-css-nested-deep/src/pages/pageA.js
@@ -1,0 +1,4 @@
+import Foo from '../components/Bar';
+import Bar from '../components/Foo';
+
+console.log('>> Page A: ', { Foo, Bar });