+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change log
 
+## 3.4.7 (2024-01-09)
+
+- fix: if the same CSS file is imported in many js files, then the CSS is extracted for the first issuer only, #68
+
 ## 3.4.6 (2024-01-02)
 
-fix: the `pathData.chunk.name` is undefined when the `js.filename` is a function, this bug was introduced in `3.4.5`
+- fix: the `pathData.chunk.name` is undefined when the `js.filename` is a function, this bug was introduced in `3.4.5`
 
 ## 3.4.5 (2024-01-01)
 
-fix: the `pathData.filename` is undefined after changes when the `js.filename` is a function
+- fix: the `pathData.filename` is undefined after changes when the `js.filename` is a function
 
 ## 3.4.4 (2023-12-28)